@@ -1,16 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<LangVersion>preview</LangVersion>
+		<TargetFramework>net8.0</TargetFramework>
 		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-		<Version>0.1.0</Version>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<InternalsVisibleTo Include="Diahon.VersionControl.Subversion.Test" />
+	</ItemGroup>
+
+	<PropertyGroup Label="Nuget">
+		<Version>0.1.0-alpha</Version>
+		<PackageDescription>Minimal cross-platform subversion client for dotnet</PackageDescription>
+		<Authors>ShortDevelopment</Authors>
+		<Company>Diahon</Company>
+		<RepositoryUrl>https://github.com/Diahon/Diahon.VersionControl.Subversion</RepositoryUrl>
+		<PackageReadmeFile>README.md</PackageReadmeFile>
+		<PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
+	</PropertyGroup>
+
+	<ItemGroup Label="Nuget">
+		<Content Include="../README.md">
+			<PackagePath>/</PackagePath>
+			<Pack>true</Pack>
+		</Content>
+		<Content Include="../LICENSE.txt">
+			<PackagePath>/</PackagePath>
+			<Pack>true</Pack>
+		</Content>
 	</ItemGroup>
 
 </Project>
